Fix nightly PDFs (#19)
The git reset --hard in commit() is (probably) there so avoid a race condition when multiple workers are building and pushing PDFs. However, if it's a situation where a PDF needs to be updated (as opposed to the creation of a new file; as is the case for nightlies), that actually wipes out the updated PDF.

With this patch, we build all the PDFs into a temporary directory and then copy them over to JULIA_DOCS only after we have run git reset --hard and updated the repository.